+++ src/test/scala/scorex/waves/WavesAPISpecification.scala
@@ -9,7 +9,7 @@ class WavesAPISpecification extends FunSuite with Matchers {
 
   import TestingCommons._
 
-  test("/waves/external-payment API route can not send to address from another net") {
+  test("/waves/broadcast-signed-payment API route can not send to address from another net") {
 
     val senderPublicKey = "GvXeYd2iFJUNV7KgeGV2cdnScyrEvrr9uPYJeQFtvg21"
     val recipient = "3PBWXDFUc86N2EQxKJmW8eFco65xTyMZx6J"
@@ -18,18 +18,18 @@ class WavesAPISpecification extends FunSuite with Matchers {
     val payment = ExternalPayment(timestamp, amount, 1L, senderPublicKey, recipient, "1")
     val json = Json.toJson(payment).toString
 
-    val response = postRequest(us = "/waves/external-payment", body = json)
+    val response = postRequest(us = "/waves/broadcast-signed-payment", body = json)
     assert(response.toString == InvalidAddress.json.toString)
   }
 
-  test("/wave/external-payment can not send from suspendedAddresses") {
+  test("/wave/broadcast-signed-payment can not send from suspendedAddresses") {
     val senderPublicKey = "5JiSMVVvadkKt2K9dDJjiwLiDzuCMmzcHnNuEzct2LiY"
     val recipient = "3N5jhcA7R98AUN12ee9pB7unvnAKfzb3nen"
     val timestamp = 1465391445252L
     val amount = 10000000000000L
     val payment = ExternalPayment(timestamp, amount, 1L, senderPublicKey, recipient, "1")
     val json = Json.toJson(payment).toString
-    val response = postRequest(us = "/waves/external-payment", body = json)
+    val response = postRequest(us = "/waves/broadcast-signed-payment", body = json)
     assert(response.toString == InvalidSender.json.toString)
   }